requirements.txt
Flask
PyMuPDF
python-docx
docx2pdf; platform_system == "Windows"
reportlab
